Move incremental merkle tree tests to zcash-gtest.
